requirements.txt
alembic==1.13.2
asyncpg==0.29.0
asyncio==3.4.3
fastapi==0.114.0
pydantic==2.9.0
SQLAlchemy==2.0.34
uvicorn==0.30.6
python-dotenv==1.0.1